Decouple PDO connection and query feature. #322
